+++ main.js
@@ -1,6 +1,7 @@
 import { dataFormat } from "./formatDate.js";
-import { getComments, postComment, login, token, setToken } from "./api.js";
+import { getComments, postComment,token } from "./api.js";
 import { renderComments } from "./renderComments.js";
+import {loginAutoriz} from "./autoriz.js";
 
 
 const buttonElement = document.querySelector('.add-form-button');
@@ -50,69 +51,69 @@ const disableButton = () => {
 nameElement.addEventListener("input", disableButton);
 textElement.addEventListener("input", disableButton);
 
-const loginElement = document.getElementById("login");
+// const loginElement = document.getElementById("login");
 
-const loginHtml = `<a class="login-link" href="#" id="link-to-in">Чтобы добавить комментарий, авторизуйтесь </a>`;
+// const loginHtml = `<a class="login-link" href="#" id="link-to-in">Чтобы добавить комментарий, авторизуйтесь </a>`;
 
-loginElement.innerHTML = loginHtml;
-
-
-
-const loginAutoriz = () => {
-    let loginLinkButton = document.getElementById("link-to-in");
-    loginLinkButton.addEventListener('click', () => {
-        const autorizationFormHtml = `
-        <div class="add-form"> Форма входа
-            <br>
-            <input id="login-input" type="text" class="add-form-name add-login" placeholder="Логин" />
-            <br>
-            <input id="pass-input" type="text" class="add-form-name add-pass" placeholder="Пароль" />
-            <button id="add-form-login" class="add-form-button">Войти</button>
-            <a class="login-link" href="#" id="link-to-login">Зарегистрироваться</a>
-        </div>`;
-
-        loginElement.innerHTML = autorizationFormHtml;
-
-        const loginButton = document.getElementById("add-form-login");
-        const loginInputElement = document.getElementById("login-input");
-        const passwordInputElement = document.getElementById("pass-input");
-
-        loginButton.addEventListener('click', () => {
-            login({
-                login: loginInputElement.value,
-                password: passwordInputElement.value,
-            }).then((responseData) => {
-                setToken(responseData.user.token);
-                nameElement.value = responseData.user.name;
-                loginElement.innerHTML = ` `;
-            });
-
-        });
+// loginElement.innerHTML = loginHtml;
 
 
-        const registrationLinkButton = document.getElementById("link-to-login");
-        registrationLinkButton.addEventListener('click', () => {
 
-            const registrationFormHtml = `
-                <div class="add-form"> Форма регистрации
-                    <input type="text" class="add-form-name add-login" placeholder="Имя" />
-                    <br>
-                    <input type="text" class="add-form-name add-login" placeholder="Логин" />
-                    <br>
-                    <input type="text" class="add-form-name add-pass" placeholder="Пароль" />
-                    <button class="add-form-button">Зарегистрироваться</button>
-                    <a class="login-link" href="#" id="link-to-in">Войти</a>
-                </div>`;
+// const loginAutoriz = () => {
+//     let loginLinkButton = document.getElementById("link-to-in");
+//     loginLinkButton.addEventListener('click', () => {
+//         const autorizationFormHtml = `
+//         <div class="add-form"> Форма входа
+//             <br>
+//             <input id="login-input" type="text" class="add-form-name add-login" placeholder="Логин" />
+//             <br>
+//             <input id="pass-input" type="text" class="add-form-name add-pass" placeholder="Пароль" />
+//             <button id="add-form-login" class="add-form-button">Войти</button>
+//             <a class="login-link" href="#" id="link-to-login">Зарегистрироваться</a>
+//         </div>`;
 
-            loginElement.innerHTML = registrationFormHtml;
-            loginAutoriz();
-        });
+//         loginElement.innerHTML = autorizationFormHtml;
 
-    });
+//         const loginButton = document.getElementById("add-form-login");
+//         const loginInputElement = document.getElementById("login-input");
+//         const passwordInputElement = document.getElementById("pass-input");
 
-};
+//         loginButton.addEventListener('click', () => {
+//             login({
+//                 login: loginInputElement.value,
+//                 password: passwordInputElement.value,
+//             }).then((responseData) => {
+//                 setToken(responseData.user.token);
+//                 nameElement.value = responseData.user.name;
+//                 loginElement.innerHTML = ` `;
+//             });
 
-loginAutoriz();
+//         });
+
+
+//         const registrationLinkButton = document.getElementById("link-to-login");
+//         registrationLinkButton.addEventListener('click', () => {
+
+//             const registrationFormHtml = `
+//                 <div class="add-form"> Форма регистрации
+//                     <input type="text" class="add-form-name add-login" placeholder="Имя" />
+//                     <br>
+//                     <input type="text" class="add-form-name add-login" placeholder="Логин" />
+//                     <br>
+//                     <input type="text" class="add-form-name add-pass" placeholder="Пароль" />
+//                     <button class="add-form-button">Зарегистрироваться</button>
+//                     <a class="login-link" href="#" id="link-to-in">Войти</a>
+//                 </div>`;
+
+//             loginElement.innerHTML = registrationFormHtml;
+//             loginAutoriz();
+//         });
+
+//     });
+
+// };
+
+loginAutoriz({nameElement});
 
 // Форма добавления нового комментария в ленту
 buttonElement.addEventListener('click', () => {
